% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_fact.R
\name{random_fact}
\alias{random_fact}
\title{Random fact}
\usage{
random_fact()
}
\value{
Two paragraphs of a random wikipedia page
}
\description{
This function provides the first or the first two paragraphs of a random wikipedia page.
}
